+++ builddir/CMakeFiles/LibXml2.dir/encoding.c.o.d
@@ -69,9 +69,10 @@ CMakeFiles/LibXml2.dir/encoding.c.o: \
  /usr/include/asm-generic/errno.h /usr/include/asm-generic/errno-base.h \
  /home/jessie/Documents/libxml2/include/libxml/encoding.h \
  /home/jessie/Documents/libxml2/include/libxml/xmlerror.h \
+ /home/jessie/Documents/libxml2/include/libxml/xmlstring.h \
+ /usr/lib/gcc/x86_64-linux-gnu/15/include/stdarg.h \
  /home/jessie/Documents/libxml2/include/libxml/xmlmemory.h \
- /usr/include/stdio.h /usr/lib/gcc/x86_64-linux-gnu/15/include/stdarg.h \
- /usr/include/x86_64-linux-gnu/bits/types/__fpos_t.h \
+ /usr/include/stdio.h /usr/include/x86_64-linux-gnu/bits/types/__fpos_t.h \
  /usr/include/x86_64-linux-gnu/bits/types/__mbstate_t.h \
  /usr/include/x86_64-linux-gnu/bits/types/__fpos64_t.h \
  /usr/include/x86_64-linux-gnu/bits/types/__FILE.h \
@@ -84,7 +85,6 @@ CMakeFiles/LibXml2.dir/encoding.c.o: \
  /usr/include/x86_64-linux-gnu/bits/stdio2.h \
  /home/jessie/Documents/libxml2/include/libxml/parser.h \
  /home/jessie/Documents/libxml2/include/libxml/tree.h \
- /home/jessie/Documents/libxml2/include/libxml/xmlstring.h \
  /home/jessie/Documents/libxml2/include/libxml/xmlregexp.h \
  /home/jessie/Documents/libxml2/include/libxml/dict.h \
  /home/jessie/Documents/libxml2/include/libxml/hash.h \