+++ builddir/CMakeFiles/LibXml2.dir/xmlIO.c.o.d
@@ -87,6 +87,7 @@ CMakeFiles/LibXml2.dir/xmlIO.c.o: /home/jessie/Documents/libxml2/xmlIO.c \
  /usr/include/x86_64-linux-gnu/bits/stdio2.h \
  /home/jessie/Documents/libxml2/include/libxml/encoding.h \
  /home/jessie/Documents/libxml2/include/libxml/xmlerror.h \
+ /home/jessie/Documents/libxml2/include/libxml/xmlstring.h \
  /home/jessie/Documents/libxml2/include/libxml/tree.h \
  /usr/lib/gcc/x86_64-linux-gnu/15/include/limits.h \
  /usr/lib/gcc/x86_64-linux-gnu/15/include/syslimits.h \
@@ -96,7 +97,6 @@ CMakeFiles/LibXml2.dir/xmlIO.c.o: /home/jessie/Documents/libxml2/xmlIO.c \
  /usr/include/x86_64-linux-gnu/bits/pthread_stack_min-dynamic.h \
  /usr/include/x86_64-linux-gnu/bits/pthread_stack_min.h \
  /usr/include/x86_64-linux-gnu/bits/posix2_lim.h \
- /home/jessie/Documents/libxml2/include/libxml/xmlstring.h \
  /home/jessie/Documents/libxml2/include/libxml/xmlmemory.h \
  /home/jessie/Documents/libxml2/include/libxml/xmlregexp.h \
  /home/jessie/Documents/libxml2/include/libxml/uri.h \