+++ builddir/CMakeFiles/LibXml2.dir/xmlmemory.c.o.d
@@ -73,6 +73,7 @@ CMakeFiles/LibXml2.dir/xmlmemory.c.o: \
  /usr/include/x86_64-linux-gnu/bits/stdio.h \
  /usr/include/x86_64-linux-gnu/bits/stdio2.h \
  /home/jessie/Documents/libxml2/include/libxml/xmlerror.h \
+ /home/jessie/Documents/libxml2/include/libxml/xmlstring.h \
  /home/jessie/Documents/libxml2/include/libxml/parser.h \
  /home/jessie/Documents/libxml2/include/libxml/tree.h \
  /usr/lib/gcc/x86_64-linux-gnu/15/include/limits.h \
@@ -83,7 +84,6 @@ CMakeFiles/LibXml2.dir/xmlmemory.c.o: \
  /usr/include/x86_64-linux-gnu/bits/pthread_stack_min-dynamic.h \
  /usr/include/x86_64-linux-gnu/bits/pthread_stack_min.h \
  /usr/include/x86_64-linux-gnu/bits/posix2_lim.h \
- /home/jessie/Documents/libxml2/include/libxml/xmlstring.h \
  /home/jessie/Documents/libxml2/include/libxml/xmlregexp.h \
  /home/jessie/Documents/libxml2/include/libxml/dict.h \
  /home/jessie/Documents/libxml2/include/libxml/hash.h \